requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-whois==0.9.4
apscheduler==3.10.4
aiosmtplib==3.0.2
pydantic==2.10.3
pydantic-settings==2.7.0
jinja2==3.1.4
aiosqlite==0.20.0
